test: split agent loop white-box suites
